move rest code code inset to workbench out of (monaco) editor
